+++ src/core/game-states/title-screen-state.cpp
@@ -43,9 +43,7 @@ void TitleScreenState::update(float deltatime) {
 	m_ui->GetRenderer()->RenderOffscreen();
 
 	// Need to restore the GPU state because noesis changes it
-	GLCall(glBindFramebuffer(GL_FRAMEBUFFER, 0));
-	GLCall(glViewport(0, 0, WIN_WIDTH, WIN_HEIGHT));
-	GLCall(glClearStencil(0));
+	restoreGpuState();
 
 	// Render
 	m_ui->GetRenderer()->Render();