+++ src/entity-factories/mirror-factory.cpp
@@ -30,7 +30,7 @@ unsigned int MirrorFactory::create(float posX, float posY) {
 	m_registry.assign<entityTag::Mirror>(myEntity);
 	m_registry.assign<cmpt::Sprite>(myEntity, m_mirrorSprite);
 	m_registry.assign<renderTag::Single>(myEntity);
-	m_registry.assign<cmpt::Transform>(myEntity, glm::vec2(posX, posY), glm::vec2(1.0f));
+	m_registry.assign<cmpt::Transform>(myEntity, glm::vec2(posX, posY), zIndexTower);
 	m_registry.assign<cmpt::Hitbox>(myEntity, MIRROR_RADIUS);
 	m_registry.assign<cmpt::ConstrainedRotation>(myEntity, 16);
 	return myEntity;