+++ src/entity-factories/tile-factory.cpp
@@ -27,14 +27,14 @@ TileFactory::~TileFactory() {
 unsigned int TileFactory::createSpawn(glm::vec2 position) {
 	auto myEntity = m_registry.create();
 	m_registry.assign<cmpt::Primitive>(myEntity, m_spawn);
-	m_registry.assign<cmpt::Transform>(myEntity, position);
+	m_registry.assign<cmpt::Transform>(myEntity, position, zIndexMap);
 	return myEntity;
 }
 
 unsigned int TileFactory::createArrival(glm::vec2 position) {
 	auto myEntity = m_registry.create();
 	m_registry.assign<cmpt::Primitive>(myEntity, m_arrival);
-	m_registry.assign<cmpt::Transform>(myEntity, position);
+	m_registry.assign<cmpt::Transform>(myEntity, position, zIndexMap);
 	return myEntity;
 }
 
@@ -43,7 +43,7 @@ unsigned int TileFactory::createPath(glm::vec2 position) {
 	m_registry.assign<cmpt::Primitive>(myEntity, m_path);
 	//m_registry.assign<cmpt::Sprite>(myEntity, m_path);
 	//m_registry.assign<renderTag::Single>(myEntity);
-	m_registry.assign<cmpt::Transform>(myEntity, position);
+	m_registry.assign<cmpt::Transform>(myEntity, position, zIndexMap);
 	return myEntity;
 }
 
@@ -52,7 +52,7 @@ unsigned int TileFactory::createConstructible(glm::vec2 position) {
 	m_registry.assign<cmpt::Primitive>(myEntity, m_constructible);
 	//m_registry.assign<cmpt::Sprite>(myEntity, m_constructible);
 	//m_registry.assign<renderTag::Single>(myEntity);
-	m_registry.assign<cmpt::Transform>(myEntity, position);
+	m_registry.assign<cmpt::Transform>(myEntity, position, zIndexMap);
 	m_registry.assign<tileTag::Constructible>(myEntity);
 	return myEntity;
 }
@@ -60,6 +60,6 @@ unsigned int TileFactory::createConstructible(glm::vec2 position) {
 unsigned int TileFactory::createLocked(glm::vec2 position) {
 	auto myEntity = m_registry.create();
 	m_registry.assign<cmpt::Primitive>(myEntity, m_locked);
-	m_registry.assign<cmpt::Transform>(myEntity, position);
+	m_registry.assign<cmpt::Transform>(myEntity, position, zIndexMap);
 	return myEntity;
 }